moved ColumnEvaluator
